webserver/requirements.txt
fastapi 
uvicorn 
cassandra-driver
